wimax:  Relax assumption of current state when receiving a ranging response
The changeset 12906:98ce80ed995f fixed a bug in the assert, but the check
for == SS_STATE_WAITING_RNG_RSP appears to be too strict, as the device
may have moved to another state.  Review of wimax examples and tests shows
that this device may be in states 4, 5, 6, 7, or 8 when this message
is received, so the assert condition was relaxed.